Refactor the reading API
