--- lambda/read.ts
@@ -1,27 +1,0 @@
-const AWS = require('aws-sdk');
-const db = new AWS.DynamoDB.DocumentClient();
-const TABLE_NAME = process.env.TABLE_NAME || '';
-const PRIMARY_KEY = process.env.PRIMARY_KEY || '';
-
-export const handler = async (event: any = {}) : Promise <any> => {
-
-  const todoId = event.pathParameters.todoId;
-  if (!todoId) {
-    return { statusCode: 400, body: `Error: You are missing the path parameter id` };
-  }
-
-  const params = {
-    TableName: TABLE_NAME,
-    Key: {
-      [PRIMARY_KEY]: todoId
-    }
-  };
-
-  try {
-    const response = await db.get(params).promise();
-    console.log(response);
-    return { statusCode: 200, body: JSON.stringify(response.Item) };
-  } catch (dbError) {
-    return { statusCode: 500, body: JSON.stringify(dbError) };
-  }
-};
